n2/render: add explicit `disable_inlining` query arg
Switch to sqlalchemy.pool.NullPool, which doesn't reuse connections,
which is purportedly fine for file-based databases.

We should probably _actually_ fix the connection leak, but it'll be
easier to rewrite the code to either use the SQLAlchemy API, or switch
to a different framework.

TODO: Convert `disable_inlining` into some kind of all-in-one "make
saveable" query arg that also disables the FlaskDebugToolbar.