note-domains: Sort by sort_time instead of time_scope_id
This has the effect of putting domains-without-sort_times at the bottom,
but those are few (for our dataset) and so we don't care.

To do this "properly", we should probably have an SQLAlchemy
hybrid_property that lets us sort rows based on a combination of
sort_time and time_scope_id, but that's probably expensive and also
those are rare occurrences.