t3/update: when updating TaskLinkages, index by time_scope
We were previously indexing by SQLAlchemy object identity, which… we
didn't understand very well. So ue the tl_ts instead, which we know
exactly the expected semantics of.

NB This situation seems to happen when modifying the `import_source` of
an existing Task, followed by all of its TaskLinkages disappearing.